Split out mongodb settings
